fix(learn): harden shared course experience
