+++ apps/learn/messages/en.json
@@ -73,16 +73,44 @@
     "empty": "Nothing here yet.",
     "learner": "Learner",
     "loading": "Loading",
+    "next": "Next",
+    "previous": "Previous",
+    "retry": "Retry",
     "save": "Save",
     "saving": "Saving"
   },
   "courses": {
+    "backToModules": "Back to modules",
+    "courseOutline": "Course outline",
+    "done": "Done",
     "empty": "Your teacher has not assigned courses yet.",
+    "extraReading": "Extra reading",
+    "flashcardCount": "{count, plural, one {# flashcard} other {# flashcards}}",
+    "flashcards": "Flashcards",
+    "left": "Left",
+    "loadError": "We could not load this course. Try again in a moment.",
     "locked": "Locked",
     "mapDescription": "Follow the course path, unlock each checkpoint, and keep every lesson visible at a glance.",
+    "moduleContent": "Module content",
+    "moduleEmpty": "This module has no content yet.",
+    "moduleHint": "Work through the content, then return to pick another module.",
+    "moduleNavigation": "Module navigation",
+    "modulePosition": "Module {current} of {total}",
+    "moduleStatus": "Module status",
     "modules": "Modules",
+    "openModule": "Open module",
+    "publishedModules": "{count, plural, one {# published module} other {# published modules}}",
+    "quizCount": "{count, plural, one {# quiz} other {# quizzes}}",
+    "quizSets": "Quiz sets",
+    "quizzes": "Quizzes",
+    "selectModule": "Select a module to view its content",
+    "sharedCourse": "Shared course",
     "title": "Courses",
-    "untitled": "Course"
+    "untitled": "Course",
+    "videoCount": "{count, plural, one {# video} other {# videos}}",
+    "videos": "Videos",
+    "xp": "XP",
+    "youtubeVideo": "YouTube video"
   },
   "home": {
     "assignments": "Due assignments",
