+++ apps/learn/messages/vi.json
@@ -73,16 +73,44 @@
     "empty": "Chưa có nội dung.",
     "learner": "Học viên",
     "loading": "Đang tải",
+    "next": "Tiếp",
+    "previous": "Trước",
+    "retry": "Thử lại",
     "save": "Lưu",
     "saving": "Đang lưu"
   },
   "courses": {
+    "backToModules": "Quay lại danh sách bài học",
+    "courseOutline": "Đề cương khóa học",
+    "done": "Xong",
     "empty": "Giáo viên chưa giao khóa học nào.",
+    "extraReading": "Đọc thêm",
+    "flashcardCount": "{count} thẻ ghi nhớ",
+    "flashcards": "Thẻ ghi nhớ",
+    "left": "Còn lại",
+    "loadError": "Không tải được khóa học này. Hãy thử lại sau giây lát.",
     "locked": "Đã khóa",
     "mapDescription": "Đi theo lộ trình khóa học, mở từng điểm mốc và nhìn rõ bài nào cần học tiếp.",
+    "moduleContent": "Nội dung bài học",
+    "moduleEmpty": "Bài học này chưa có nội dung.",
+    "moduleHint": "Học xong nội dung rồi quay lại chọn bài khác.",
+    "moduleNavigation": "Điều hướng bài học",
+    "modulePosition": "Bài {current} trong {total}",
+    "moduleStatus": "Trạng thái bài học",
     "modules": "Bài học",
+    "openModule": "Mở bài học",
+    "publishedModules": "{count} bài học đã xuất bản",
+    "quizCount": "{count} câu hỏi",
+    "quizSets": "Bộ câu hỏi",
+    "quizzes": "Câu hỏi",
+    "selectModule": "Chọn một bài để xem nội dung",
+    "sharedCourse": "Khóa học được chia sẻ",
     "title": "Khóa học",
-    "untitled": "Khóa học"
+    "untitled": "Khóa học",
+    "videoCount": "{count} video",
+    "videos": "Video",
+    "xp": "XP",
+    "youtubeVideo": "Video YouTube"
   },
   "home": {
     "assignments": "Bài tập cần làm",
